Block nodes: use ref callback
